Back up and Restore script first issue
First commit of the working backup and restore scripts

each one takes a volume parameter and a file parameter

The backup writes the data from the specified volume into a tar.gz file

The restore takes the tar.gz file and dumps it into the specified volume.
If the specified volume exists the user is prompted to confirm they are OK
to delete the volume and recreate it.

Actually the current script always asks if it is OK to destroy, not only if
it exists. Will need to update that in the future.